How to put this politely.... I2c is being really fcking annoying >:)

+++ src/Source/Orientation.cpp
@@ -238,7 +238,7 @@ namespace Orientation
         return compassHeading;
     }
 
-    double getYaw()
+    float getYaw()
     {
         return yaw;
     }
